requirements.txt
fastapi
uvicorn
twilio
openai
python-decouple
sqlalchemy
psycopg2-binary
python-multipart
pyngrok
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas